Create Github action to create images

+++ meteor-tools/meteor-node-runtime/scripts/startup.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 set -e -x
 
-BUNDLE_FILE="/bundle/bundle.tar.gz"
-APP_DIR="$HOME/app"
+BUNDLE_FILE=${BUNDLE_FILE:-"/bundle/bundle.tar.gz"}
+APP_DIR=${APP_DIR:-"$HOME/app"}
+NO_NPM_INSTALL=${NO_NPM_INSTALL}
 
 if [ -f "${BUNDLE_FILE}" ]; then
   mkdir "$APP_DIR"
@@ -10,7 +11,7 @@ if [ -f "${BUNDLE_FILE}" ]; then
   tar -xzf "${BUNDLE_FILE}" -C "$APP_DIR" --strip 1 --group="$GROUP" --owner="$USER"
 fi
 
-if [ -f "${APP_DIR}/programs/server/package.json" ]; then
+if [ -f "${APP_DIR}/programs/server/package.json" ] && [ "x${NO_NPM_INSTALL}" == "x" ] ; then
   cd "${APP_DIR}/programs/server/" && npm install --unsafe-perm && cd -
 fi
